refactor: move localStorage mock helper to neutral hooks home
