+++ apps/web/hooks/use-local-storage-boolean.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, beforeEach, afterEach, vi } from "vitest";
 import { act, renderHook, waitFor } from "@testing-library/react";
 import { useLocalStorageBoolean } from "./use-local-storage-boolean";
-import { makeLocalStorageMock } from "@/hooks/domains/integrations/local-storage-mock.test-helpers";
+import { makeLocalStorageMock } from "@/hooks/local-storage-mock.test-helpers";
 
 const STORAGE_KEY = "kandev:generic:flag:v1";
 const SYNC_EVENT = "kandev:generic:flag-changed";
